Main folder structure change
