+++ Cavokator/Cavokator.csproj
@@ -153,34 +153,36 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AboutFragment.cs" />
-    <Compile Include="AirportConverter.cs" />
-    <Compile Include="AirportDefinitions.cs" />
-    <Compile Include="ApplyTheme.cs" />
-    <Compile Include="ChangelogDialog.cs" />
-    <Compile Include="ClickableSpan.cs" />
-    <Compile Include="NotamContainer.cs" />
-    <Compile Include="NotamDialogMap.cs" />
-    <Compile Include="NotamDialogRaw.cs" />
-    <Compile Include="NotamFetcher.cs" />
-    <Compile Include="NotamFragment.cs" />
-    <Compile Include="NotamOptionsDialog.cs" />
-    <Compile Include="NotamTypeD.cs" />
-    <Compile Include="NotamTypeQ.cs" />
-    <Compile Include="SettingsFragment.cs" />
-    <Compile Include="WxColorCoder.cs" />
-    <Compile Include="ConditionFragment.cs" />
-    <Compile Include="WxFetcher.cs" />
-    <Compile Include="WxContainer.cs" />
+    <Compile Include="cavokator.config\AboutFragment.cs" />
+    <Compile Include="cavokator.notam\NotamCardsAdapter.cs" />
+    <Compile Include="cavokator.util\AirportConverter.cs" />
+    <Compile Include="cavokator.util\AirportDefinitions.cs" />
+    <Compile Include="cavokator.config\ApplyTheme.cs" />
+    <Compile Include="cavokator.config\ChangelogDialog.cs" />
+    <Compile Include="cavokator.util\ClickableSpan.cs" />
+    <Compile Include="cavokator.util\NestedScrollViewListener.cs" />
+    <Compile Include="cavokator.notam\NotamContainer.cs" />
+    <Compile Include="cavokator.notam\NotamDialogMap.cs" />
+    <Compile Include="cavokator.notam\NotamDialogRaw.cs" />
+    <Compile Include="cavokator.notam\NotamFetcher.cs" />
+    <Compile Include="cavokator.notam\NotamFragment.cs" />
+    <Compile Include="cavokator.notam\NotamOptionsDialog.cs" />
+    <Compile Include="cavokator.notam\NotamTypeD.cs" />
+    <Compile Include="cavokator.notam\NotamTypeQ.cs" />
+    <Compile Include="cavokator.config\SettingsFragment.cs" />
+    <Compile Include="cavokator.weather\WxColorCoder.cs" />
+    <Compile Include="cavokator.condition\ConditionFragment.cs" />
+    <Compile Include="cavokator.weather\WxFetcher.cs" />
+    <Compile Include="cavokator.weather\WxContainer.cs" />
     <Compile Include="Resources\Resource.Designer.cs" />
     <Compile Include="ActivityWxMain.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="WxMetarFragment.cs" />
-    <Compile Include="WxOptionsDialog.cs" />
-    <Compile Include="ConditionDecoder.cs" />
-    <Compile Include="ConditionDialog.cs" />
-    <Compile Include="ConditionContainer.cs" />
-    <Compile Include="WxTaforDivider.cs" />
+    <Compile Include="cavokator.weather\WxMetarFragment.cs" />
+    <Compile Include="cavokator.weather\WxOptionsDialog.cs" />
+    <Compile Include="cavokator.condition\ConditionDecoder.cs" />
+    <Compile Include="cavokator.condition\ConditionDialog.cs" />
+    <Compile Include="cavokator.condition\ConditionContainer.cs" />
+    <Compile Include="cavokator.weather\WxTaforDivider.cs" />
   </ItemGroup>
   <ItemGroup>
     <AndroidAsset Include="Assets\appicon.psd" />
@@ -585,6 +587,7 @@
   <ItemGroup>
     <AndroidResource Include="Resources\layout\notam_airports.xml" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\Android\Xamarin.Android.CSharp.targets" />
   <Import Project="..\packages\Xamarin.Android.Support.Annotations.26.1.0.1\build\MonoAndroid80\Xamarin.Android.Support.Annotations.targets" Condition="Exists('..\packages\Xamarin.Android.Support.Annotations.26.1.0.1\build\MonoAndroid80\Xamarin.Android.Support.Annotations.targets')" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
