+++ Cavokator/cavokator.condition/ConditionContainer.cs
