+++ Cavokator/cavokator.condition/ConditionDecoder.cs
