+++ Cavokator/cavokator.condition/ConditionDialog.cs
