+++ Cavokator/cavokator.condition/ConditionFragment.cs
