+++ Cavokator/cavokator.config/ChangelogDialog.cs
