+++ Cavokator/cavokator.config/SettingsFragment.cs
