+++ Cavokator/cavokator.notam/NotamDialogMap.cs
