+++ Cavokator/cavokator.notam/NotamOptionsDialog.cs
