+++ Cavokator/cavokator.notam/NotamTypeQ.cs
