+++ Cavokator/cavokator.util/AirportDefinitions.cs
