+++ Cavokator/cavokator.util/ClickableSpan.cs
