+++ Cavokator/cavokator.util/NestedScrollViewListener.cs
@@ -1,0 +1,33 @@
+﻿using System;
+using Android.Support.V4.Widget;
+
+namespace Cavokator
+{
+    internal sealed class NestedScrollViewListener : Java.Lang.Object, NestedScrollView.IOnScrollChangeListener
+    {
+        public EventHandler<NestedScrollViewListenerEventArgs> OnScrollEvent;
+        
+        public void OnScrollChange(NestedScrollView v, int scrollX, int scrollY, int oldScrollX, int oldScrollY)
+        {
+            Console.WriteLine("scrollY = " + scrollY);
+            Console.WriteLine("scrollY = " + oldScrollY);
+
+            ScrollEvent(scrollY);
+        }
+
+        private void ScrollEvent(int y)
+        {
+            OnScrollEvent?.Invoke(this, new NestedScrollViewListenerEventArgs(y));
+        }
+    }
+
+    internal class NestedScrollViewListenerEventArgs
+    {
+        public int PositionY { get; }
+
+        public NestedScrollViewListenerEventArgs(int y)
+        {
+            PositionY = y;
+        }
+    }
+}
