+++ Cavokator/cavokator.weather/WxColorCoder.cs
