+++ Cavokator/cavokator.weather/WxOptionsDialog.cs
